[OPIK-1421] feedback scores editor
